# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RoleMesh API
  version: 1.1.0
  description: |
    REST contract for the `/api/v1/*` surface introduced by the
    webui-backend v1.1 design. Only `/api/v1/backends` is implemented
    in Phase 0; the rest are listed as forward contracts so frontend
    work can compile against typed types from day one.

    The companion WebSocket endpoint `/api/v1/conversations/{id}/stream`
    is documented in design §4 (event-driven JSON protocol); OpenAPI 3
    does not model WS, so it is intentionally omitted here.

    Error responses use the shape declared in design §13:
    `{code, message, details?}`.
servers:
  - url: /
    description: Same-origin (FastAPI serves both the SPA and the API).
...
paths:
...
  /api/v1/models:
    get:
      tags: [Models]
      operationId: listModels
      summary: List platform models
      description: |
        Public, tenant-agnostic catalog. Optional `provider` /
        `family` query filters narrow the picker for the coworker
        wizard. `is_active=false` rows are filtered out by default
        — admin surfaces are out of scope for v1.1 (design §14).
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: query
          name: provider
          required: false
          schema: { $ref: '#/components/schemas/ModelProvider' }
        - in: query
          name: family
          required: false
          schema: { $ref: '#/components/schemas/ModelFamily' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Model' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/models/{id}:
    parameters:
      - $ref: '#/components/parameters/IdInPath'
    get:
      tags: [Models]
      operationId: getModel
      summary: Fetch one platform model
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Model' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/admin/models:
    post:
      tags: [Admin]
      operationId: adminModelsCreate
      summary: Register a new platform model (admin only)
      description: |
        The `models` catalog is platform-managed — tenants pick from
        it but only operators add rows. POST/PATCH/DELETE require
        the `owner` role (design §13 role check).
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ModelCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Model' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /api/v1/admin/models/{id}:
    patch:
      tags: [Admin]
      operationId: adminModelsUpdate
      summary: Update a platform model (admin only)
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/IdInPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ModelUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Model' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
    delete:
      tags: [Admin]
      operationId: adminModelsDelete
      summary: Retire a platform model (admin only)
      description: |
        Soft delete — sets `active = false` rather than dropping the
        row so existing coworkers that point at the model keep
        rendering. Returns 409 if any coworker is still bound.
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/IdInPath' }
      responses:
        '204':
          description: Deleted (soft)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdInPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Caller is authenticated but not permitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: |
        Resource conflict — typically a duplicate name within the
        tenant, or attempted mutation of a terminal-state run.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unprocessable:
      description: |
        Semantic validation failure (e.g. `BACKEND_INCOMPAT`,
        `MISSING_CREDENTIAL` — design §13).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Stable, machine-readable code. Known values include
            `RESOURCE_IN_USE`, `BACKEND_INCOMPAT`, `MISSING_CREDENTIAL`,
            `BOOTSTRAP_NEEDS_TENANT_OWNER`, `NEEDS_USER_PRESENCE`,
            `WS_TICKET_EXPIRED`, `SKILL_MANIFEST_PROTECTED`,
            `REAUTH_REQUIRED` (design §13).
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    ModelProvider:
      type: string
      enum: [anthropic, bedrock, openai, google]
    ModelFamily:
      type: string
      enum: [claude, gpt, gemini, llama]
...
    Model:
      type: object
      required: [id, provider, model_id, model_family, display_name, is_active]
      properties:
        id: { type: string, format: uuid }
        provider: { $ref: '#/components/schemas/ModelProvider' }
        model_id:
          type: string
          description: |
            Provider-side model identifier (e.g. `claude-opus-4-7`).
        model_family: { $ref: '#/components/schemas/ModelFamily' }
        display_name: { type: string }
        is_active: { type: boolean }
        created_at:
          type: string
          format: date-time
          nullable: true
...
    ModelCreate:
      type: object
      required: [provider, model_id, model_family, display_name]
      properties:
        provider: { $ref: '#/components/schemas/ModelProvider' }
        model_id:
          type: string
          minLength: 1
          maxLength: 200
        model_family: { $ref: '#/components/schemas/ModelFamily' }
        display_name:
          type: string
          minLength: 1
          maxLength: 200
        is_active:
          type: boolean
          default: true
    ModelUpdate:
      type: object
      description: |
        All fields optional. `is_active: false` here is preferred
        over DELETE for already-bound models (DELETE returns 409 on
        in-use rows; this lets an operator hide a deprecated model
        from new bindings without dropping the row).
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 200
        is_active:
          type: boolean

    # -----------------------------------------------------------------
    # WebSocket frame schemas (PR23).
    #
    # OpenAPI 3 has no native channel/subscription model (that's
    # AsyncAPI's job), but the JSON SHAPE of every frame in a ws
    # message is plain JSON, so it fits cleanly under components/
    # schemas. Each direction is a discriminated union on the `type`
    # field; codegen produces a tagged union on the TS side and the
    # backend uses Pydantic discriminated unions for symmetric
    # validation (see `WsServerEventModel` in schemas_v1.py).
    #
    # Channel: WS /api/v1/conversations/{id}/stream
    # (handshake authenticated via `?ticket=` query, minted by
    # POST /auth/ws-ticket).
    # -----------------------------------------------------------------
...